+++ CHANGELOG.md
@@ -5,9 +5,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [2.1.3] - Unreleased
 
 ### Changed
+
 ### Added
+- Add `clientonly` script to start only the electron client for a remote server.
 - Add symbol and color properties of event when `CALENDAR_EVENTS` notification is broadcasted from `default/calendar` module.
+
 ### Updated
+
 ### Fixed
 
 - Fixed issue with incorrect allignment of analog clock when displayed in the center column of the MM