+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.1.3] - Unreleased
+
+### Changed
+### Added
+### Updated
+### Fixed
+
 ## [2.1.2] - 2017-07-01
 
 ### Changed